Allow option of using order-sign hook
Allow the option of using either the order-select or order-sign hook to trigger calling the CDS Service to determine if the order meets the acceptable use guidelines. Using the order-sign hook would seem to reduce the needed number of calls to the CDS Service since all the order details are complete at this point. With using the order-select hook, the CDS Service would need to be called whenever the order details are changed since a change in the details may change whether the radiology order meets the acceptable use criteria.